fixup! genesys: usbhub: Fix firmware size for GL3590 hubs

+++ plugins/genesys/fu-genesys-usbhub-firmware.c
@@ -69,7 +69,7 @@ fu_genesys_usbhub_firmware_parse(FuFirmware *firmware,
 	gsize bufsz = 0;
 	const guint8 *buf = g_bytes_get_data(fw, &bufsz);
 	guint8 sign[4];
-	guint16 code_size = 0x6000;
+	guint32 code_size = 0x6000;
 	guint16 version_raw = 0;
 	gboolean is3590 = FALSE;
 	g_autofree gchar *version = NULL;
